+++ lib/memory/allocator.c
@@ -19,17 +19,17 @@ static inline void cu_CAllocator_Free(void *self, cu_Slice mem) {
   free(raw);
 }
 
-static cu_Slice_Result cu_CAllocator_Alloc(
-    void *self, size_t size, size_t alignment) {
+static cu_Slice_Result cu_CAllocator_Alloc(void *self, cu_Layout layout) {
   CU_UNUSED(self);
-  if (size == 0) {
+  if (layout.elem_size == 0) {
     cu_Io_Error err = {
         .kind = CU_IO_ERROR_KIND_INVALID_INPUT, .errnum = Size_Optional_none()};
     return cu_Slice_result_error(err);
   }
 
-  if (alignment < sizeof(void *))
-    alignment = sizeof(void *);
+  size_t size = layout.elem_size;
+  size_t alignment = layout.alignment;
+  if (alignment < sizeof(void *)) alignment = sizeof(void *);
 
   size_t total = size + alignment - 1 + sizeof(void *);
   void *raw = malloc(total);
@@ -48,17 +48,20 @@ static cu_Slice_Result cu_CAllocator_Alloc(
 }
 
 static cu_Slice_Result cu_CAllocator_Resize(
-    void *self, cu_Slice mem, size_t size, size_t alignment) {
+    void *self, cu_Slice mem, cu_Layout layout) {
   CU_UNUSED(self);
-  CU_IF_NULL(mem.ptr) { return cu_CAllocator_Alloc(self, size, alignment); }
-  if (size == 0) {
+  CU_IF_NULL(mem.ptr) { return cu_CAllocator_Alloc(self, layout); }
+  if (layout.elem_size == 0) {
     cu_CAllocator_Free(NULL, mem);
     cu_Io_Error err = {
         .kind = CU_IO_ERROR_KIND_INVALID_INPUT, .errnum = Size_Optional_none()};
     return cu_Slice_result_error(err);
   }
 
-  cu_Slice_Result new_mem = cu_CAllocator_Alloc(self, size, alignment);
+  size_t size = layout.elem_size;
+  size_t alignment = layout.alignment;
+
+  cu_Slice_Result new_mem = cu_CAllocator_Alloc(self, layout);
   if (!cu_Slice_result_is_ok(&new_mem)) {
     return new_mem;
   }
@@ -82,22 +85,19 @@ cu_Allocator cu_Allocator_CAllocator(void) {
 
 CU_OPTIONAL_IMPL(cu_Allocator, cu_Allocator)
 
-static cu_Slice_Result cu_null_alloc(
-    void *self, size_t size, size_t alignment) {
+static cu_Slice_Result cu_null_alloc(void *self, cu_Layout layout) {
   CU_UNUSED(self);
-  CU_UNUSED(size);
-  CU_UNUSED(alignment);
+  CU_UNUSED(layout);
   cu_Io_Error err = {
       .kind = CU_IO_ERROR_KIND_OUT_OF_MEMORY, .errnum = Size_Optional_none()};
   return cu_Slice_result_error(err);
 }
 
 static cu_Slice_Result cu_null_resize(
-    void *self, cu_Slice mem, size_t size, size_t alignment) {
+    void *self, cu_Slice mem, cu_Layout layout) {
   CU_UNUSED(self);
   CU_UNUSED(mem);
-  CU_UNUSED(size);
-  CU_UNUSED(alignment);
+  CU_UNUSED(layout);
   cu_Io_Error err = {
       .kind = CU_IO_ERROR_KIND_OUT_OF_MEMORY, .errnum = Size_Optional_none()};
   return cu_Slice_result_error(err);