+++ tests/test_allocator.cpp
@@ -18,7 +18,8 @@ TEST(Allocator, GPABasic) {
   cfg.bucketSize = 64;
   cu_Allocator alloc = cu_Allocator_GPAllocator(&gpa, cfg);
 
-  cu_Slice_Result mem_res = cu_Allocator_Alloc(alloc, 32, 8);
+  cu_Slice_Result mem_res =
+      cu_Allocator_Alloc(alloc, cu_Layout_create(32, 8));
   ASSERT_TRUE(cu_Slice_result_is_ok(&mem_res));
   cu_Slice mem = mem_res.value;
   cu_Memory_memset(mem.ptr, 0xAA, mem.length);